@@ -1,57 +1,53 @@
 {
   "name": "smith-app",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "pwd && ng build --prod --base-href=smith-app",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "update": "ng update --all"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.5",
-    "@angular/cdk": "^6.2.1",
-    "@angular/common": "^6.0.5",
-    "@angular/compiler": "^6.0.5",
-    "@angular/core": "^6.0.5",
-    "@angular/forms": "^6.0.5",
-    "@angular/http": "^6.0.5",
-    "@angular/material": "^6.2.1",
-    "@angular/platform-browser": "^6.0.5",
-    "@angular/platform-browser-dynamic": "^6.0.5",
-    "@angular/router": "^6.0.5",
-    "core-js": "^2.5.7",
+    "@angular/animations": "^8.2.3",
+    "@angular/cdk": "^8.1.3",
+    "@angular/common": "~8.2.3",
+    "@angular/compiler": "~8.2.3",
+    "@angular/core": "~8.2.3",
+    "@angular/forms": "~8.2.3",
+    "@angular/material": "^8.1.3",
+    "@angular/platform-browser": "~8.2.3",
+    "@angular/platform-browser-dynamic": "~8.2.3",
+    "@angular/router": "~8.2.3",
     "d3": "^5.4.0",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.2.1",
-    "sprintf-js": "^1.1.1",
-    "zone.js": "^0.8.26"
+    "rxjs": "~6.4.0",
+    "sprintf-js": "^1.1.2",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^8.3.0",
-    "@angular/compiler-cli": "^6.0.5",
-    "@angular/language-service": "^6.0.5",
-    "@types/d3": "^5.0.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~10.3.3",
-    "@types/sprintf-js": "^1.1.0",
-    "codelyzer": "^4.3.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "~8.3.0",
+    "@angular/compiler-cli": "~8.2.3",
+    "@angular/language-service": "~8.2.3",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "@types/sprintf": "^0.1.0",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.2",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.0",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^3.5.3"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
