@@ -38,7 +38,7 @@ export class MarkerDetailsComponent implements OnInit, OnDestroy {
   tableDataSource = new MatTableDataSource(this.data);
   dataSourceInput: DataSource<Element> | Observable<Element[]> | Element[] | null = this.data;
 
-  @ViewChild(MatTable) matTable: MatTable<Element>;
+  @ViewChild(MatTable, { static: false }) matTable: MatTable<Element>;
 
   constructor(private route: ActivatedRoute, private router: Router, private state: StateService) {
 
